+++ test/unit/test_wiz.py
@@ -16,6 +16,12 @@ import wiz.exception
 
 
 @pytest.fixture()
+def mocked_registry_defaults(mocker):
+    """Return mocked 'wiz.registry.get_default' function."""
+    return mocker.patch.object(wiz.registry, "get_defaults")
+
+
+@pytest.fixture()
 def mocked_fetch_definition_mapping(mocker):
     """Return mocked 'wiz.fetch_definition_mapping' function."""
     return mocker.patch.object(wiz, "fetch_definition_mapping")
@@ -207,9 +213,9 @@ def test_fetch_package_request_from_command_error():
     "with-environ",
 ])
 def test_resolve_context(
+    mocked_registry_defaults, mocked_fetch_definition_mapping,
     mocked_graph_resolver, mocked_package_initiate_environ,
-    mocked_package_extract_context, mocked_utility_encode,
-    mocker, options
+    mocked_package_extract_context, mocked_utility_encode, mocker, options
 ):
     """Get resolved context mapping."""
     requests = ["test1 >=10, < 11", "test2", "test3[variant]"]
@@ -246,6 +252,73 @@ def test_resolve_context(
         "registries": paths
     }
 
+    mocked_registry_defaults.assert_not_called()
+    mocked_fetch_definition_mapping.assert_not_called()
+
+    mocked_graph_resolver.assert_called_once_with("__PACKAGE_DEFINITIONS__")
+    mocked_resolver.compute_packages.assert_called_once_with([
+        Requirement(request) for request in requests
+    ])
+
+    mocked_package_initiate_environ.assert_called_once_with(
+        options.get("environ_mapping")
+    )
+
+    mocked_package_extract_context.assert_called_once_with(
+        packages, environ_mapping="__INITIAL_ENVIRON__"
+    )
+
+
+@pytest.mark.parametrize("options", [
+    {},
+    {"environ_mapping": "__ENVIRON__"},
+], ids=[
+    "without-environ",
+    "with-environ",
+])
+def test_resolve_context_with_default_definition_mapping(
+    mocked_registry_defaults, mocked_fetch_definition_mapping,
+    mocked_graph_resolver, mocked_package_initiate_environ,
+    mocked_package_extract_context, mocked_utility_encode, mocker, options
+):
+    """Get resolved context mapping with default definition mapping."""
+    requests = ["test1 >=10, < 11", "test2", "test3[variant]"]
+    paths = ["/path/to/registry1", "/path/to/registry2"]
+
+    context = {"environ": {"KEY": "VALUE"}, "command": {"app": "APP"}}
+    packages = [
+        mocker.Mock(identifier="test1"),
+        mocker.Mock(identifier="test2"),
+        mocker.Mock(identifier="test3")
+    ]
+
+    mocked_resolver = mocker.Mock(**{"compute_packages.return_value": packages})
+    mocked_graph_resolver.return_value = mocked_resolver
+    mocked_package_initiate_environ.return_value = "__INITIAL_ENVIRON__"
+    mocked_package_extract_context.return_value = context
+    mocked_utility_encode.return_value = "__ENCODED_CONTEXT__"
+    mocked_registry_defaults.return_value = paths
+    mocked_fetch_definition_mapping.return_value = {
+        "package": "__PACKAGE_DEFINITIONS__",
+        "registries": paths
+    }
+
+    result = wiz.resolve_context(requests, **options)
+
+    assert result == {
+        "environ": {
+            "KEY": "VALUE",
+            "WIZ_VERSION": __version__,
+            "WIZ_CONTEXT": "__ENCODED_CONTEXT__"
+        },
+        "command": {"app": "APP"},
+        "packages": packages,
+        "registries": paths
+    }
+
+    mocked_registry_defaults.asset_called_once()
+    mocked_fetch_definition_mapping.asset_called_once_with(paths)
+
     mocked_graph_resolver.assert_called_once_with("__PACKAGE_DEFINITIONS__")
     mocked_resolver.compute_packages.assert_called_once_with([
         Requirement(request) for request in requests
@@ -268,9 +341,9 @@ def test_resolve_context(
     "with-environ",
 ])
 def test_resolve_context_with_implicit_packages(
+    mocked_registry_defaults, mocked_fetch_definition_mapping,
     mocked_graph_resolver, mocked_package_initiate_environ,
-    mocked_package_extract_context, mocked_utility_encode,
-    mocker, options
+    mocked_package_extract_context, mocked_utility_encode, mocker, options
 ):
     """Get resolved context mapping with implicit packages."""
     requests = ["test1 >=10, < 11", "test2", "test3[variant]"]
@@ -309,6 +382,9 @@ def test_resolve_context_with_implicit_packages(
         "registries": paths
     }
 
+    mocked_registry_defaults.asset_called_once()
+    mocked_fetch_definition_mapping.asset_called_once_with(paths)
+
     mocked_graph_resolver.assert_called_once_with("__PACKAGE_DEFINITIONS__")
     mocked_resolver.compute_packages.assert_called_once_with([
         Requirement(request) for request in requests + implicit
@@ -331,9 +407,9 @@ def test_resolve_context_with_implicit_packages(
     "with-environ",
 ])
 def test_resolve_context_with_implicit_packages_ignored(
+    mocked_registry_defaults, mocked_fetch_definition_mapping,
     mocked_graph_resolver, mocked_package_initiate_environ,
-    mocked_package_extract_context, mocked_utility_encode,
-    mocker, options
+    mocked_package_extract_context, mocked_utility_encode, mocker, options
 ):
     """Get resolved context mapping with implicit packages ignored."""
     requests = ["test1 >=10, < 11", "test2", "test3[variant]"]
@@ -372,6 +448,9 @@ def test_resolve_context_with_implicit_packages_ignored(
         "packages": packages,
         "registries": paths
     }
+
+    mocked_registry_defaults.asset_called_once()
+    mocked_fetch_definition_mapping.asset_called_once_with(paths)
 
     mocked_graph_resolver.assert_called_once_with("__PACKAGE_DEFINITIONS__")
     mocked_resolver.compute_packages.assert_called_once_with([