+++ doc/release/release_notes.rst
@@ -4,6 +4,15 @@
 Release Notes
 *************
 
+.. release:: Upcoming
+
+    .. change:: changed
+        :tags: API
+
+        Updated :func:`wiz.resolve_context` to make the *definition_mapping*
+        argument optional. If no definition mapping is provided, a sensible one
+        will be fetched from default registries.
+
 .. release:: 0.15.1
     :date: 2018-08-14