+++ doc/release/release_notes.rst
@@ -13,6 +13,48 @@ Release Notes
         argument optional. If no definition mapping is provided, a sensible one
         will be fetched from default registries.
 
+    .. change:: changed
+        :tags: API
+
+        Updated :meth:`~wiz.definition.Definition.remove` method to return
+        :class:`~wiz.definition.Definition` instance without raising
+        :exc:`KeyError` exception when element to remove does not exist.
+
+    .. change:: changed
+        :tags: API
+
+        Updated :meth:`~wiz.definition.Definition.remove_key` method to return
+        :class:`~wiz.definition.Definition` instance without raising
+        :exc:`KeyError` exception when element to remove does not exist.
+
+    .. change:: changed
+        :tags: API
+
+        Updated :meth:`~wiz.definition.Definition.remove_key` method to return
+        copy of a :class:`~wiz.definition.Definition` instance without element
+        mapping if the latest key is removed.
+
+    .. change:: changed
+        :tags: API
+
+        Updated :meth:`~wiz.definition.Definition.remove_index` method to return
+        :class:`~wiz.definition.Definition` instance without raising
+        :exc:`KeyError` exception when index to remove does not exist.
+
+    .. change:: changed
+        :tags: API
+
+        Updated :meth:`~wiz.definition.Definition.remove_index` method to return
+        copy of a :class:`~wiz.definition.Definition` instance without element
+        list if the latest item is removed.
+
+    .. change:: fixed
+        :tags: API
+
+        Fixed :mod:`wiz.mapping` to prevent serialisation of boolean values as
+        it causes validation errors when serialized mapping is used to create
+        a new :class:`~wiz.definition.Definition` instance.
+
 .. release:: 0.15.1
     :date: 2018-08-14